quantization_base -> quantized_tensor rename
