+++ tests/pytorch/attention/test_attention.py
@@ -45,7 +45,7 @@ from transformer_engine.pytorch.utils import (
 )
 from transformer_engine.pytorch.utils import get_cudnn_version
 import transformer_engine_torch as tex
-from transformer_engine.pytorch.quantization_base import (
+from transformer_engine.pytorch.quantized_tensor import (
     Quantizer,
     prepare_for_saving,
     restore_from_saved,
