+++ transformer_engine/debug/pytorch/debug_quantization.py
@@ -15,7 +15,7 @@ import torch
 import transformer_engine_torch as tex
 
 from transformer_engine.common.recipe import Recipe
-from transformer_engine.pytorch.quantization_base import (
+from transformer_engine.pytorch.quantized_tensor import (
     QuantizedTensor,
     Quantizer,
     QuantizedTensorStorage,
