+++ transformer_engine/pytorch/ops/_common.py
@@ -13,7 +13,7 @@ from transformer_engine_torch import FP8TensorMeta
 from .. import torch_version
 from ..quantization import FP8GlobalStateManager
 from ..tensor.float8_tensor import Float8Tensor
-from ..quantization_base import QuantizedTensorStorage
+from ..quantized_tensor import QuantizedTensorStorage
 from ..utils import canonicalize_dtype
