+++ transformer_engine/pytorch/tensor/float8_blockwise_tensor.py
@@ -14,7 +14,7 @@ from transformer_engine_torch import Float8BlockScaleTensorFormat
 
 from transformer_engine.common.recipe import Float8BlockScaling, Recipe
 from .storage.float8_blockwise_tensor_storage import Float8BlockwiseQTensorStorage
-from ..quantization_base import QuantizedTensor, Quantizer
+from ..quantized_tensor import QuantizedTensor, Quantizer
 from ._quantization_helpers import _IdentityFunc
 from ..utils import devices_match, round_up_to_nearest_multiple
