+++ transformer_engine/pytorch/tensor/float8_tensor.py
@@ -14,7 +14,7 @@ from transformer_engine_torch import DType as TE_DType
 from transformer_engine.common.recipe import DelayedScaling, Float8CurrentScaling, Recipe
 from ..utils import canonicalize_process_group, devices_match
 from .storage.float8_tensor_storage import Float8TensorStorage, _FromFloat8Func
-from ..quantization_base import QuantizedTensor, Quantizer
+from ..quantized_tensor import QuantizedTensor, Quantizer
 from ._quantization_helpers import _IdentityFunc
 from ..constants import dist_group_type
