+++ transformer_engine/pytorch/tensor/storage/float8_blockwise_tensor_storage.py
@@ -13,7 +13,7 @@ import transformer_engine_torch as tex
 from transformer_engine_torch import DType as TE_DType
 from transformer_engine_torch import Float8BlockScaleTensorFormat
 
-from ...quantization_base import QuantizedTensorStorage, Quantizer
+from ...quantized_tensor import QuantizedTensorStorage, Quantizer
 
 from ...constants import TE_DType_To_Torch
