+++ transformer_engine/pytorch/tensor/storage/float8_tensor_storage.py
@@ -12,7 +12,7 @@ import torch
 import transformer_engine_torch as tex
 from transformer_engine_torch import DType as TE_DType
 
-from ...quantization_base import QuantizedTensorStorage, Quantizer
+from ...quantized_tensor import QuantizedTensorStorage, Quantizer
 
 from ...constants import TE_DType as torch_to_transformer_engine_dtype
