+++ transformer_engine/pytorch/tensor/storage/nvfp4_tensor_storage.py
@@ -16,7 +16,7 @@ import torch
 # import transformer_engine_torch as tex
 from transformer_engine_torch import DType as TE_DType
 
-from ...quantization_base import QuantizedTensorStorage, Quantizer
+from ...quantized_tensor import QuantizedTensorStorage, Quantizer
 
 # from ...constants import TE_DType as torch_to_transformer_engine_dtype
 from ...utils import _empty_tensor
