+++ transformer_engine/pytorch/utils.py
@@ -12,7 +12,7 @@ import numpy as np
 import torch
 
 from . import torch_version
-from .quantization_base import Quantizer
+from .quantized_tensor import Quantizer
 from ..debug.pytorch.debug_quantization import DebugQuantizedTensor
